Add shell function example
This example shows my intended usage. When this file is sourced it allows to use:
`cd2repo git@github.com:quidome/git2url.git`

Which will clone the repository to `$HOME/Workspace/github.com/quidome/git2url` if that path does not exist on the system yet and then changes the working dir to it.